requirements.txt
amqp==2.1.4
appdirs==1.4.0
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.1
Babel==2.3.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
cached-property==1.3.0
certifi==2017.1.23
channels==1.0.2
constantly==15.1.0
daphne==1.0.1
defusedxml==0.4.1
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
flower==0.9.1
futures==3.0.5
incremental==16.10.1
isodate==0.5.4
msgpack-python==0.4.8
osa==0.2
psycopg2
PyJWT==1.4.2
pytz==2016.10
redis==2.10.5
requests==2.13.0
requests-toolbelt==0.7.0
six==1.10.0
suds==0.4
tornado==4.2
Twisted==16.6.0
txaio==2.6.0
vine==1.1.3
zope.interface==4.3.3
# windows only requirement
# pypiwin32==219
